+++ scraper/scraper.ts
@@ -26,10 +26,12 @@ import {
   scrapeDurhamRegion,
   scrapeBrantford,
   scrapePeterborough,
+  scrapeSmithsFalls,
 } from './engines/custom';
 import { scrapeRSS } from './engines/rss';
 import { scrapeJazzHR } from './engines/jazzhhr';
 import { scrapeWorkland } from './engines/workland';
+import { scrapeJibe } from './engines/jibe';
 
 export { scrapeSuccessFactors, scrapeWorkday, scrapeWaterfront, scrapeConservationHalton, scrapeADP };
 export { urlId, scrapeRawAndStage } from './utils';
@@ -90,11 +92,21 @@ async function main() {
   await scrapeBambooHR(db, context, 'https://cityofhamilton.bamboohr.com/careers', 'City of Hamilton');
   await scrapePeterborough(db, context);
 
-  // 7. Eastern Ontario
+  // 7. Southwestern Ontario
+  await scrapeJazzHR(db, context, 'https://cityofwindsor.applytojob.com/apply/', 'City of Windsor', 'windsor');
+  await scrapeADP(db, context, 'https://workforcenow.adp.com/mascsr/default/mdf/recruitment/recruitment.html?cid=9ba4d624-1cab-4482-861f-900704c3df0d&ccId=19000101_000001&lang=en_CA', 'City of Sarnia');
+  await scrapeDayforce(db, context, 'https://jobs.dayforcehcm.com/en-CA/stthomas/CANDIDATEPORTAL', 'City of St. Thomas');
+  await scrapeJobs2Web(db, context, 'https://careers.regionofwaterloo.ca/RoW/search/', 'Region of Waterloo');
+
+  // 8. Northern Ontario
+  await scrapeJibe(db, context, 'https://careers.thunderbay.ca/careers-home/jobs', 'City of Thunder Bay', 'thunderbay');
+
+  // 9. Eastern Ontario
   await scrapeSuccessFactors(db, context, 'https://career47.sapsf.com/careers/cityofottawa/search', 'City of Ottawa', 'https://career47.sapsf.com');
   await scrapeRSS(db, context, 'https://careers.cityofkingston.ca/CL2/net/ResumeProcessing/RssFeedOutput.aspx?CLID=61577&lang=1', 'City of Kingston', 'kingston');
   await scrapeJazzHR(db, context, 'https://cityofbelleville.applytojob.com/apply/', 'City of Belleville', 'belleville');
   await scrapeWorkland(db, context, 'https://atlas.workland.com/careers/cornwall/jobs?page=1', 'City of Cornwall', 'cornwall');
+  await scrapeSmithsFalls(db, context);
 
   console.log('\nCleaning up expired jobs...');
   await cleanupExpiredJobs(db, runStartedAt);